Seperated Queue into its own file
