+++ src/Test/MainTests.cpp
@@ -4,6 +4,7 @@
 #include <boost/test/included/unit_test.hpp>
 
 #include "../Data Structures/data_structures.hpp"
+#include "../Data Structures/queue.hpp"
 #include "../Tokeniser/tokeniser.hpp"
 #include <string>
